Move with_term test utility to _tokenizer
