+++ src/_tokenizer/names.rs
@@ -83,7 +83,7 @@ pub fn match_zwj(text: &str) -> Option<((), &str)> {
 
 #[cfg(test)]
 mod tests {
-    use crate::tests::{return_none, with_term};
+    use crate::_tokenizer::tests::{return_none, with_term};
     use rstest::rstest;
     use std::str::FromStr;
