+++ src/_tokenizer/punctuators.rs
@@ -80,7 +80,7 @@ pub fn match_right_brace_punctuator(text: &str) -> Option<((), &str)> {
 
 #[cfg(test)]
 mod tests {
-    use crate::tests::{return_none, with_term};
+    use crate::_tokenizer::tests::{return_none, with_term};
     use rstest::rstest;
     use std::str::FromStr;
